sergio/ejercicio01/AppEngine/requirements.txt
Flask==1.0.2
google-cloud-storage==1.9.0
oauth2client==4.1.3
GoogleAppEngineCloudStorageClient==1.9.22.1
